Set a timeout on "Do you want"-prompts
If we need too long to respond gpg thinks it's agent died and tries to
restart it. This doesn't work with split-gpg and is anyway not what we
want.

This also has the advantage that if a scripts tries to to a bunch of
operations without waiting for success all the windows can be cleaned up
by just waiting for the timeout.